requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
entrypoints==0.3
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
Pillow==7.1.2
pycairo==1.19.1
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4.1
pyxdg==0.25
qrcode==6.1
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
spidev==3.5
ssh-import-id==5.7
urllib3==1.24.1
